feat: firebase auth first implement
